requirements.txt
Django==4.2.13
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
pymongo==4.6.0
python-dotenv==1.0.0
